+++ code/backend/weibo-service/src/main/java/com/ilife/weiboservice/repository/UserRepository.java
@@ -9,7 +9,7 @@ import javax.transaction.Transactional;
 public interface UserRepository extends CrudRepository<User, Integer> {
 
     //这里findById会报错，所以改了一下
-    User findAllById(Integer id);
+    User findAllById(Long id);
 
     User findByNickname(String nickname);
